Added description for extractpubkey.py
script that supports the export of Public Key for algorithm types: RSA, DSA, and EC. By exporting the public key, you can easily integrate it with other systems or applications that need to verify signatures or encrypt data meant for the owner of the private key. The public key also plays a crucial role in generating digital certificates. When an entity wants to obtain a digital certificate, they send a request to a Certificate Authority (CA). This request includes the public key, which serves as proof of identity. The CA verifies the request and issues a certificate that links the public key to the identity of the certificate holder.

This certificate can be used for various applications, such as:
   * SSL/TLS encryption to protect website connections.
   * Encrypted and digitally signed emails.
   * Verification of an individual's or organization's identity in online transactions.