Script for Windows and Linux using HSM and PCKS11
Creates or uses an exsisting an RSA Key Pair on an nShield HSM device 
Wraps the RSA private key with an exisiting or newly generated AES wrapping key. 
File is exported in encrypted binary format.
Generates a CSR using the RSA key stored in HSM via OpenSSL with nfkm engine.